fix: :adhesive_bandage: Carga del modelo 3d
@Dannkol

+++ app/Modificar/routes.py
@@ -196,9 +196,11 @@ def actualizar_producto(producto_id):
             else:
                 cursor.execute("SELECT Asset FROM Productos WHERE Id = %s", (producto_id,))
                 resultado = cursor.fetchone()
-                model_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..' ,'assets',resultado[0])
-                print(model_path)
-                os.remove(model_path)
+                if resultado[0] != None :
+                    print('data: ',resultado)
+                    model_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..' ,'assets',resultado[0])
+                    print(model_path)
+                    os.remove(model_path)
                 asset_path = "models/{}".format(asset)
                 cursor.execute(
                     "UPDATE Productos SET Nombre = %s, Precio = %s, Descripcion = %s, Ref = %s, Asset = %s WHERE Id = %s",
